requirements.txt
Django==3.2.7  
django-rest-framework==0.1.0
djangorestframework==3.12.4
psycopg2==2.9.1
requests==2.27.1
beautifulsoup4==4.10.0
lxml==4.7.1